# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "GoRide API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/auth/google/callback": {
            "get": {
                "description": "Handles Google OAuth2 callback",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Google Callback",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/auth/google/login": {
            "get": {
                "description": "Initiates Google OAuth2 login",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Google Auth",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v1/auth/login": {
            "post": {
                "description": "Login a user and return user details and token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login a user",
                "parameters": [
                    {
                        "description": "User login data",
                        "name": "loginDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login successful",
                        "schema": {
                            "$ref": "#/definitions/dto.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request, validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error, database or service errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v1/auth/register": {
            "post": {
                "description": "Register a new user and assign roles if provided.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user with roles",
                "parameters": [
                    {
                        "description": "User registration data",
                        "name": "registerDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UserRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registration successful",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad request, validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error, database or service errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.UserLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "Q2Sb9@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "Pass123!@"
                }
            }
        },
        "dto.UserLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/dto.UserResponse"
                }
            }
        },
        "dto.UserRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "password_confirm",
                "roles",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "Q2Sb9@example.com"
                },
                "password": {
                    "type": "string",
                    "minLength": 8,
                    "example": "Cilok99!@"
                },
                "password_confirm": {
                    "type": "string",
                    "example": "Cilok99!@"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user"
                    ]
                },
                "username": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3,
                    "example": "John Doe"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Roles    string `json:\"role\" db:\"role\"` // e.g., \"user\", \"driver\", \"admin\"\nAvatarUrl  sql.NullString `json:\"avatar_url\" db:\"avatar_url\"`\nAvatarName sql.NullString `json:\"avatar_name\" db:\"avatar_name\"`\nFirstName  sql.NullString `json:\"first_name\" db:\"first_name\"`\nLastName   sql.NullString `json:\"last_name\" db:\"last_name\"`\nPhone      sql.NullString `json:\"phone\" db:\"phone\"`\nAddress    sql.NullString `json:\"address\" db:\"address\"`",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "Optional, for soft delete",
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "picture": {
                    "description": "URL to the user's profile picture",
                    "type": "string"
                },
                "provider": {
                    "description": "e.g., \"google\", \"local\"",
                    "type": "string"
                },
                "provider_id": {
                    "description": "ID from the provider (e.g., Google ID)",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}